fix(mcp): split tool search registration schema
