+++ open-sse/mcp-server/server.ts
@@ -6,11 +6,10 @@ import {
   getComboStepTarget,
 } from "../../src/lib/combos/steps.ts";
 
-import { handleToolSearch } from "./toolSearch/handler.ts";
+import { registerToolSearchTool } from "./toolSearch/register.ts";
 
 import {
   MCP_TOOLS,
-  toolSearchInput,
   getHealthInput,
   listCombosInput,
   getComboMetricsInput,
@@ -1201,21 +1200,7 @@ export function createMcpServer(): McpServer {
     )
   );
 
-  server.registerTool(
-    "omniroute_tool_search",
-    {
-      description:
-        "Search MCP tools by keyword; returns compact one-line TS signatures for token-efficient discovery.",
-      inputSchema: toolSearchInput,
-    },
-    withScopeEnforcement("omniroute_tool_search", (args) => {
-      const parsed = toolSearchInput.parse(args ?? {});
-      const result = handleToolSearch(parsed);
-      return Promise.resolve({
-        content: [{ type: "text" as const, text: JSON.stringify(result, null, 2) }],
-      });
-    })
-  );
+  registerToolSearchTool(server, withScopeEnforcement);
 
   // ── Memory Tools ──────────────────────────────
   Object.values(memoryTools).forEach((toolDef: any) => {
